Rename polling to upstream module
