+++ tests/polling/test_polling_service.py
@@ -1,7 +1,7 @@
 import pytest
 from unittest.mock import Mock, patch
-from hutch_bunny.core.polling.polling_service import PollingService
-from hutch_bunny.core.polling.task_api_client import TaskApiClient
+from hutch_bunny.core.upstream.polling_service import PollingService
+from hutch_bunny.core.upstream.task_api_client import TaskApiClient
 from logging import Logger
 import requests
